moved ContinuousNode.py to continuous.factors, moved tests too
