+++ pgmpy/continuous/factors/ContinuousNode.py
