+++ pgmpy/tests/test_continuous/test_ContinuousNode.py
